Bring components into a single file
