--- frontend/apps/picasso/components/Organisms/Static/PrivacyPolicy/ApplicableLawAndJurisdiction.tsx
@@ -1,12 +1,0 @@
-import { NormalText } from "@/components/Organisms/Static/NormalText";
-
-export const ApplicableLawAndJurisdiction = () => (
-  <section>
-    <NormalText>
-      This Privacy Policy shall be governed by and construed in accordance with
-      the laws of Saint Lucia, without regard to conflict of law principles. Any
-      disputes arising in respect of this Privacy Policy shall be submitted to
-      the exclusive jurisdiction of the courts of Saint Lucia.
-    </NormalText>
-  </section>
-);
