--- frontend/apps/picasso/components/Organisms/Static/PrivacyPolicy/DataSecurity.tsx
@@ -1,15 +1,0 @@
-import { NormalText } from "@/components/Organisms/Static/NormalText";
-
-export const DataSecurity = () => (
-  <section>
-    <NormalText>
-      We implement and maintain reasonable administrative, physical, and
-      technical security safeguards to help protect your Personal Data from
-      outside attacks or threats, loss, theft, misuse, unauthorized access,
-      disclosure, alteration, and destruction. Nevertheless, we do not provide
-      any guarantee regarding the effectiveness of these protective measures or
-      our ability to prevent other parties, acting unlawfully, from gaining
-      access to your Personal Data.
-    </NormalText>
-  </section>
-);
