--- frontend/apps/picasso/components/Organisms/Static/PrivacyPolicy/InternationalTransfer.tsx
@@ -1,12 +1,0 @@
-export const InternationalTransfer = () => (
-  <section>
-    Information collected through the Services may be transferred to, processed,
-    stored, and used in the European Economic Area (“EEA”), the UK, and other
-    jurisdictions. Data protection laws in the EEA, the UK and other
-    jurisdictions may be different from the laws of your country of residence.
-    Your use of the Services or provision of any information therefore
-    constitutes your consent to the transfer to and from, processing, usage,
-    sharing, and storage of your Personal Data in the EEA, the UK and other
-    jurisdictions as set out in this Privacy Policy.
-  </section>
-);
