--- frontend/apps/picasso/components/Organisms/Static/PrivacyPolicy/LegalAge.tsx
@@ -1,15 +1,0 @@
-import { NormalText } from "../NormalText";
-
-export const LegalAge = () => (
-  <section>
-    <NormalText>
-      The Services are intended for general audiences and are directed to users
-      who are of legal age. Minors who lack the legal capacity under any
-      applicable laws are generally not permitted to access or use our Services.
-      Any person who becomes aware of or believes that a minor is using our
-      Services should immediately contact us. We shall remove any information
-      collected from the minor who may have used our Services without our
-      knowledge to ensure compliance with applicable laws.
-    </NormalText>
-  </section>
-);
