--- frontend/apps/picasso/components/Organisms/Static/PrivacyPolicy/OtherParties.tsx
@@ -1,25 +1,0 @@
-import { NormalText } from "@/components/Organisms/Static/NormalText";
-
-export const OtherParties = () => (
-  <section>
-    <NormalText>
-      We may integrate technologies operated or controlled by other parties into
-      parts of the Services. For example, the Services may include links that
-      hyperlink to websites, platforms, and other services not operated or
-      controlled by us.
-    </NormalText>
-    <NormalText>
-      When you interact with other parties, including when you leave the Site,
-      the other parties may independently collect information about you and
-      solicit information from you. The information collected and stored by
-      other parties remains subject to their own policies and practices,
-      including the information they share with us, your rights and choices on
-      their services and devices, and whether they store information anywhere in
-      the world. We do not control such other parties or any of their content.
-      We shall not be held liable or be responsible for the contents and privacy
-      policies of or services provided by the other parties. We encourage you to
-      read and familiarize yourself with their privacy policies and terms of
-      use.
-    </NormalText>
-  </section>
-);
