--- frontend/apps/picasso/components/Organisms/Static/PrivacyPolicy/OtherSources.tsx
@@ -1,11 +1,0 @@
-import { NormalText } from "@/components/Organisms/Static/NormalText";
-
-export const OtherSources = () => (
-  <section>
-    <NormalText>
-      We do not collect Personal Data relating to you from other sources. If we
-      do, we will inform you in advance by contacting you through the contact
-      information you have provided.
-    </NormalText>
-  </section>
-);
