--- frontend/apps/picasso/components/Organisms/Static/PrivacyPolicy/PersonalDataCollected.tsx
@@ -1,81 +1,0 @@
-import { NormalText } from "@/components/Organisms/Static/NormalText";
-import { List } from "../List";
-import { LatinListItem } from "@/components/Organisms/Static/LatinListItem";
-
-export const PersonalDataCollected = () => (
-  <section>
-    <NormalText>
-      When you access any of the Services, the Company may collect the following
-      categories of data from you or about you which may be considered as
-      Personal Data as defined in Part 1(2) of the PDPA and other privacy laws,
-      rules, and regulations as they may become applicable (the “Personal Data”
-      or “Information”):
-    </NormalText>
-    <List sx={{ listStylePosition: "outside" }}>
-      <LatinListItem>
-        Contact information, such as name, email address, telephone and/or
-        mobile phone number
-      </LatinListItem>
-      <LatinListItem>
-        Identification information, such as nationality, identification number
-        and date of birth
-      </LatinListItem>
-      <LatinListItem>
-        Image information, such as selfie, photo of national ID card
-      </LatinListItem>
-      <LatinListItem>
-        Transaction information relating to your use of the Services including,
-        without limitation, your transaction(s) with other users, deposit and
-        withdrawa
-      </LatinListItem>
-      <LatinListItem>
-        Device information, in particular as to the computer, device, or browser
-        that you rely on in order to access the Site or use the Services,
-        operating system name and version, device manufacturer and model,
-        language, internet browser type and version, and the name and version of
-        the Services you are using
-      </LatinListItem>
-      <LatinListItem>
-        Your IP address: The Company automatically logs your IP address (a
-        number that is assigned to the computer that you use to access the
-        internet) in its server log files whenever you access the Site or use
-        the Services, along with the time of your visit and the page(s) that you
-        visited.
-      </LatinListItem>
-      <LatinListItem>
-        Account information, information that is generated by your account
-        activity including, but not limited to, trading activity, order
-        activity, deposit, withdrawal, and account balances
-      </LatinListItem>
-      <LatinListItem>
-        Correspondence, information that you provide, including those we have
-        not solicited from you (in such a case, you are solely responsible for
-        such information) to use in correspondence, including during the opening
-        the account, and ongoing customer support
-      </LatinListItem>
-      <LatinListItem>
-        Publicly-available blockchain data such as your blockchain wallet
-        address, as applicable
-      </LatinListItem>
-      <LatinListItem>
-        Additional information that we may have requested to ensure
-        implementation of any contractual obligations or in connection with the
-        exercise of your right(s) under this Privacy Policy
-      </LatinListItem>
-    </List>
-    <NormalText>
-      In addition to the foregoing, you may also be required to provide
-      additional Personal Data when you, among others, require support, send us
-      information for troubleshooting or other analysis, participate in a survey
-      or other promotional activities of the Company or solicit any other
-      communication from the Company.
-    </NormalText>
-    <NormalText>
-      You may decide whether or not to submit any Personal Data required by the
-      Company. You should, however, ensure that all Personal Data submitted to
-      us is complete, accurate, true and correct. Failure on your part to do so
-      may result in our inability to provide you with the Services that you have
-      requested.
-    </NormalText>
-  </section>
-);
