--- frontend/apps/picasso/components/Organisms/Static/PrivacyPolicy/Preface.tsx
@@ -1,55 +1,0 @@
-import { NormalText } from "@/components/Organisms/Static/NormalText";
-import { Link } from "@/components";
-
-export const Preface = () => (
-  <section>
-    <NormalText>
-      This Privacy Policy, among others, seeks to notify and inform you as a
-      user (“you”, “your”, or “user”) of how Composable Finance Ltd.
-      (“Composable”, “we”, “our”, “us” or the “Company”) collects, uses, manages
-      and shares Personal Data (as hereinafter defined) in our websites at{" "}
-      <Link
-        variant="body3"
-        href="https://www.composable.finance/"
-        target="_blank"
-      >
-        https://www.composable.finance/
-      </Link>
-      ,{" "}
-      <Link variant="body3" href="https://picasso.xyz">
-        picasso.xyz
-      </Link>
-      ,{" "}
-      <Link variant="body3" href="/">
-        app.picasso.xyz
-      </Link>
-      , and such other websites, web apps, or online location that links to this
-      Privacy Policy (collectively, the “Site”) and in connection with the
-      services provided by the Company or through a related affiliate in the
-      Site (collectively, the “Services”), as well as your rights and choices
-      regarding such Personal Data, in compliance with Saint Lucia’s Privacy and
-      Data Protection Act (“PDPA”), the United Kingdom’s (“UK”) Data Protection
-      Act, the European Union’s General Data Privacy Regulation (“GDPR”) and
-      other privacy laws, rules, and regulations as they may become applicable.
-      This Privacy Policy applies to Personal Data collected, used, stored,
-      disclosed and/or processed by the Company.
-    </NormalText>
-    <NormalText>
-      By submitting information to us, or signing for the Services offered by
-      us, you agree and consent to the collection, use, retention and disclosure
-      by the Company and any of its affiliates, representatives, agents,
-      advisors and/or service providers, as well as any other activities
-      described in this Privacy Policy, of your Personal Data and other
-      information. If you do not agree with the terms of this Privacy Policy,
-      you should immediately discontinue the use of the Services and refrain
-      from accessing the Site.
-    </NormalText>
-    <NormalText>
-      We reserve the right to revise and update this Privacy Policy at any time.
-      Any changes will be effective immediately upon our posting of the latest
-      version of the Privacy Policy in the Site. Your continued use of the
-      Services indicates your consent to the latest version of the Privacy
-      Policy then published or posted.
-    </NormalText>
-  </section>
-);
