--- frontend/apps/picasso/components/Organisms/Static/PrivacyPolicy/RefusalAndNonConsent.tsx
@@ -1,12 +1,0 @@
-import { NormalText } from "@/components/Organisms/Static/NormalText";
-
-export const RefusalAndNonConsent = () => (
-  <section>
-    <NormalText>
-      If you refuse to provide your Personal Data or to consent to the transfer
-      thereof in accordance with this Privacy Policy, we may not be able to
-      fulfill our contractual requirements or, in certain cases, may not be able
-      to continue providing our Services.
-    </NormalText>
-  </section>
-);
