--- frontend/apps/picasso/components/Organisms/Static/PrivacyPolicy/RetentionPeriod.tsx
@@ -1,11 +1,0 @@
-import { NormalText } from "@/components/Organisms/Static/NormalText";
-
-export const RetentionPeriod = () => (
-  <section>
-    <NormalText>
-      We store and retain Personal Data you have provided for as long as
-      necessary for your continued use of the Services, pursuant to your
-      contract with us, and in compliance with applicable laws and regulations.
-    </NormalText>
-  </section>
-);
