--- frontend/apps/picasso/components/Organisms/Static/PrivacyPolicy/WithdrawalAndDeactivation.tsx
@@ -1,29 +1,0 @@
-import { NormalText } from "@/components/Organisms/Static/NormalText";
-
-export const WithdrawalAndDeactivation = () => (
-  <section>
-    <NormalText>
-      If you decide to stop using our Services, or deactivate your account (if
-      applicable), all Personal Data or information collected relating to you
-      shall be handled in accordance with Saint Lucia law, this Privacy Policy,
-      and the Company’s other policies (if applicable). Such deactivation shall
-      not be considered as a withdrawal of consent for the use and disclosure of
-      such Personal Data or information by us, unless otherwise expressly
-      requested or informed by you in writing.
-    </NormalText>
-    <NormalText>
-      If you decide to withdraw your consent for the Company to use and/or
-      disclose all your Personal Data, we will cease to collect your information
-      unless there is a legal justification for the continued collection of your
-      Personal Data. Further, we may not be able to continue providing our
-      Services to you or continue any contractual relationship that is in place
-      depending on the nature of your request. The withdrawal of your consent
-      may result in the termination of any agreements with us and it may be
-      considered as a breach of your contractual obligations or undertakings. In
-      such instance, we reserve our legal right to pursue any remedies available
-      at law or in equity. The withdrawal of your consent does not in any way
-      affect the lawfulness of the collection of your data based on the consent
-      given prior to the withdrawal.
-    </NormalText>
-  </section>
-);
