--- frontend/apps/picasso/components/Organisms/Static/PrivacyPolicy/YourRights.tsx
@@ -1,51 +1,0 @@
-import { NormalText } from "@/components/Organisms/Static/NormalText";
-
-export const YourRights = () => (
-  <section>
-    <NormalText>
-      Your rights under this Privacy Policy include the right to: (1) request
-      access and obtain a copy of your Personal Data; (2) rectify, correct, or
-      supplement Personal Data we have collected.; (3) object to or restrict the
-      processing of your Personal Data; (4) request portability of your Personal
-      Data; (5) object to decisions based solely on automated processing,
-      including profiling, unless you have given your consent or the same is
-      necessary for the performance of a contract between you and the Company;
-      (6) request for the removal or deletion of any of your Personal Data.
-      Additionally, if we have collected and processed your Personal Data with
-      your consent, you have the right to withdraw your consent at any time. The
-      withdrawal of your consent shall, however not affect the lawfulness of
-      processing based on your consent before its withdrawal.
-    </NormalText>
-    <NormalText>
-      Notwithstanding the foregoing, we cannot edit or delete information that
-      is stored on a particular blockchain. This information may include
-      transaction data (i.e., purchases, sales, and transfers) related to your
-      blockchain wallet address and any non-fungible tokens (NFTs) held by your
-      wallet address.
-    </NormalText>
-    <NormalText>
-      To exercise any of these rights, please contact us via our email or postal
-      address provided in this Privacy Policy and specify the right(s) you are
-      seeking to exercise. We will respond to your request within thirty (30)
-      days. We may require specific information from you to help us confirm your
-      identity and process your request. We may retain information, as
-      necessary, to fulfill the purpose for which it was collected and may
-      continue to retain and use information even after your request in
-      accordance with our legitimate interests, including as is it is necessary
-      to comply with our legal obligations, resolve disputes, prevent fraud, and
-      enforce our agreements. Notwithstanding this, you retain your right to
-      lodge a complaint with the data protection regulator in your jurisdiction.
-    </NormalText>
-    <NormalText>
-      While you can choose freely whether or not you will provide the Company
-      with your Personal Data, you may be unable to use the Services, or a
-      portion thereof, until you have provided the Company with such necessary
-      information or data.
-    </NormalText>
-    <NormalText>
-      Your data protection rights are not absolute and we may deny you your
-      rights in accordance with the applicable data protections laws by
-      providing you with the reason(s) for the denial of your request.
-    </NormalText>
-  </section>
-);
