--- frontend/apps/picasso/components/Organisms/Static/TermsOfUse/AccessAndAcceptance.tsx
@@ -1,11 +1,0 @@
-import { NormalText } from "@/components/Organisms/Static/NormalText";
-
-export const AccessAndAcceptance = () => (
-  <section>
-    <NormalText>
-      By accessing or interacting with this Site, our Services, and/or any of
-      the Composable protocols, you hereby acknowledge and accept the foregoing
-      obligations and conditions outlined in these Terms.
-    </NormalText>
-  </section>
-);
