--- frontend/apps/picasso/components/Organisms/Static/TermsOfUse/Assignment.tsx
@@ -1,13 +1,0 @@
-import { NormalText } from "@/components/Organisms/Static/NormalText";
-
-export const Assignment = () => (
-  <section>
-    <NormalText>
-      These Terms may be assigned without your prior consent to any of
-      Composable’s affiliates and its successors in the interest of any business
-      associated with the Services provided by us. You may not assign or
-      transfer any rights or obligations under this agreement without our prior
-      written consent.
-    </NormalText>
-  </section>
-);
