--- frontend/apps/picasso/components/Organisms/Static/TermsOfUse/AssumptionOfRisks.tsx
@@ -1,217 +1,0 @@
-import { NormalText } from "@/components/Organisms/Static/NormalText";
-import { List } from "../List";
-import { LatinListItem } from "@/components/Organisms/Static/LatinListItem";
-
-export const AssumptionOfRisks = () => (
-  <section>
-    <NormalText>You represent and warrant that you:</NormalText>
-    <List>
-      <LatinListItem>
-        Have the necessary technical expertise and ability to review and
-        evaluate the security, integrity and operation of any transactions that
-        you engage in through the Site or Services;
-      </LatinListItem>
-      <LatinListItem>
-        Have the knowledge, experience, understanding, professional advice and
-        information to make your own evaluation of the merits, risks and
-        applicable compliance requirements under applicable law of engaging in
-        transactions through the Site or Services;
-      </LatinListItem>
-      <LatinListItem>
-        Understand and agree to the inherent risks associated with cryptographic
-        systems and blockchain-based networks, digital assets, including the
-        usage and intricacies of native digital assets, smart contract-based
-        tokens (including fungible tokens and NFTs), and systems that interact
-        with blockchain-based networks. Composable does not own or control all
-        of the underlying software through which other blockchain networks are
-        formed. For example, the software underlying certain blockchain networks
-        is open source, such that anyone can use, copy, modify, and distribute
-        it;
-      </LatinListItem>
-      <LatinListItem>
-        Acknowledge that Composable’s underlying software and software
-        application, including the Digital Assets such as the financial NFTs
-        (“fNFTs”), are still new, unproven and in the early development stage.
-        There is an inherent risk that the software could contain weaknesses,
-        vulnerabilities, or bugs causing, inter alia, the complete loss of the
-        Digital Assets, or the risk the Digital Assets may not have their
-        intended functionalities or may have no functions at all;
-      </LatinListItem>
-      <LatinListItem>
-        Acknowledge that any use or interaction with the Services requires a
-        comprehensive understanding of applied cryptography and computer science
-        to appreciate the inherent risks, including those listed above. You
-        represent and warrant that you possess relevant knowledge and skills.
-        Any reference to a type of digital asset on the Site or otherwise during
-        the use of the Services does not indicate our approval or disapproval of
-        the technology on which the digital asset relies, and should not be used
-        as a substitute for your understanding of the risks specific to each
-        type of digital asset;
-      </LatinListItem>
-      <LatinListItem>
-        Acknowledge and understand that cryptography is a progressing field with
-        advances in code cracking or other technical advancements, such as the
-        development of quantum computers, which may present risks to digital
-        assets and the Services, and could result in the theft or loss of your
-        digital assets. To the extent possible, we intend to update
-        Composable-developed smart contracts related to the Services to account
-        for any advances in cryptography and to incorporate additional security
-        measures necessary to address risks presented from technological
-        advancements, but that intention does not guarantee or otherwise ensure
-        full security of the Services;
-      </LatinListItem>
-      <LatinListItem>
-        Acknowledge and agree that (a) Composable is not solely responsible for
-        the operation of the blockchain-based software and networks underlying
-        the Services, (b) there exists no guarantee of the functionality,
-        security, or availability of that software and networks, and (c) the
-        underlying blockchain-based networks are subject to sudden changes in
-        operating rules, such as those commonly referred to as “forks,” which
-        may materially affect the Services;
-      </LatinListItem>
-      <LatinListItem>
-        Understand that Pablo and the Picasso parachain remains under
-        development, which creates technological and security risks when using
-        the Services in addition to uncertainty relating to digital assets and
-        transactions therein. You acknowledge that the cost of transacting in
-        Pablo and the Picasso parachain is variable and may increase at any time
-        causing impact to any activities taking place therein, which may result
-        in price fluctuations or increased costs when using the Services;
-      </LatinListItem>
-      <LatinListItem>
-        Understand that blockchain networks use public and private key
-        cryptography. Thus, you alone are responsible for securing your private
-        key(s). We do not have access to your private key(s). Losing control of
-        your private key(s) will permanently and irreversibly deny you access to
-        any of your digital assets on the network. Neither Composable nor any
-        other person or entity will be able to retrieve or protect your digital
-        assets. If your private key(s) are lost, then you will not be able to
-        transfer your digital assets to any other blockchain address or wallet.
-        If this occurs, then you will not be able to realize any value or
-        utility from the digital assets that you may hold;
-      </LatinListItem>
-      <LatinListItem>
-        Understand that the markets for these digital assets are highly volatile
-        due to factors including (but not limited to) adoption, speculation,
-        technology, security, and regulation;
-      </LatinListItem>
-      <LatinListItem>
-        Confirm that you are solely responsible for your use of the Services,
-        including all of your transfers of Digital Assets and all the trades you
-        place, including any erroneous orders that may be filled. We do not take
-        any action to resolve erroneous trades or transfers that result from
-        your mistake or inadvertence;
-      </LatinListItem>
-      <LatinListItem>
-        Acknowledge that Composable’s underlying software and software
-        application are still in an early development stage and unproven. There
-        is an inherent risk that the software could contain weaknesses,
-        vulnerabilities, or bugs causing, inter alia, the complete loss of
-        digital assets and tokens.
-      </LatinListItem>
-      <LatinListItem>
-        Acknowledge that the Services are subject to flaws and that you are
-        solely responsible for evaluating any code provided by the Services or
-        Site. This warning and other warnings that Composable provides in these
-        Terms are in no way evidence or represent an on-going duty to alert you
-        to all of the potential risks of utilizing the Services or accessing the
-        Site;
-      </LatinListItem>
-      <LatinListItem>
-        Understand all the risks relating to supplying Digital Assets to a
-        liquidity to a liquidity pool including the risk of impermanent loss
-        wherein the value of one Digital Asset rises in comparison to the other
-        resulting in the assets being worth less in paper which may become
-        permanent if you withdraw liquidity before the price stabilizes;
-      </LatinListItem>
-      <LatinListItem>
-        Acknowledge and accept that these Digital Assets, especially the fNFTs,
-        are relatively new innovations and technologies in DeFi whose regulatory
-        treatment are presently uncertain, which could lead to them being
-        considered as securities and subject to restrictions under applicable
-        law concerning their purchase, sale, trade and related transactions. Any
-        unintended violation of law may lead to enforcement action and
-        penalties. Also, there is a risk with new technologies that they may not
-        function entirely as intended or create unknown risks;
-      </LatinListItem>
-      <LatinListItem>
-        Acknowledge and accept the risk that your digital assets may lose some
-        or all of their value while they are supplied to the App, you may suffer
-        large and immediate financial loss due to the fluctuation of prices of
-        Digital Assets in a liquidity pool or trading pair, and may experience
-        price slippage and cost. Thus, you should not hold value you cannot
-        afford to lose in digital assets;
-      </LatinListItem>
-      <LatinListItem>
-        Agree and accept that the Services and your digital assets could be
-        impacted by one or more regulatory inquiries or regulatory actions,
-        which could impede or limit the ability of Composable to continue to
-        make available our proprietary software and could impede or limit your
-        ability to access or use the Services;
-      </LatinListItem>
-      <LatinListItem>
-        Understand that anyone can create a token, including fake versions of
-        existing tokens and tokens that falsely claim to represent projects, and
-        acknowledge and accept the risk that you may mistakenly trade those or
-        other tokens;
-      </LatinListItem>
-      <LatinListItem>
-        Understand that digital assets and tokens may be subject to
-        expropriation and/or theft by hackers or other malicious groups by
-        obstructing the token smart contract which creates the tokens in a
-        variety of ways, including, but not limited to, malware attacks, denial
-        of service attacks, consensus-based attacks, Sybil attacks, smurfing and
-        spoofing;
-      </LatinListItem>
-      <LatinListItem>
-        Understand and accept that DEXes require no form of Know-Your-Customer
-        due diligence before users can trade and anyone with a crypto wallet can
-        trade on DEXes without any discrimination, which, thus, may potentially
-        increase the risk of interacting with malicious or fraudulent users;
-      </LatinListItem>
-      <LatinListItem>
-        Acknowledge that DEXes, because of their decentralized nature, are not
-        presently subject to comprehensive regulation;
-      </LatinListItem>
-      <LatinListItem>
-        Acknowledge and accept that the cost and speed of transacting with
-        cryptographic and blockchain-based systems are variable and may increase
-        at any time;
-      </LatinListItem>
-      <LatinListItem>
-        Understand and accept that you are solely responsible for reporting and
-        paying any taxes applicable to your use of the Services;
-      </LatinListItem>
-      <LatinListItem>
-        Confirm and accept that there are risks associated with the use of the
-        Site and Services that Composable cannot anticipate. Such risks may
-        appear as unanticipated variations or combinations of the risks
-        discussed above;
-      </LatinListItem>
-      <LatinListItem>
-        Understand and accept that Composable has the right to disable or modify
-        access to the Site and the Services (such as restricting features of the
-        Services) at any time in the event of any breach of these Terms,
-        including, if we reasonably believe any of your representations and
-        warranties may be untrue or inaccurate, and we will not be liable to you
-        for any losses or damages you may suffer as a result of or in connection
-        with the Site or the Services being inaccessible to you at any time or
-        for any reason;
-      </LatinListItem>
-      <LatinListItem>
-        Understand that the Site and the Services may evolve, which means
-        Composable may apply changes, replace, or discontinue (temporarily or
-        permanently) the Services at any time in our sole discretion; and
-      </LatinListItem>
-      <LatinListItem>
-        Assume, and agree that Composable will have no responsibility or
-        liability for any and all the risks associated with the use of the Site
-        and Services, including, but not limited to the above and you hereby
-        irrevocably waive, release and discharge all claims, whether known or
-        unknown to you, against Composable, its affiliates and their respective
-        shareholders, members, directors, officers, employees, agents and
-        representatives related to any of the risks set forth herein.
-      </LatinListItem>
-    </List>
-  </section>
-);
