--- frontend/apps/picasso/components/Organisms/Static/TermsOfUse/Eligibility.tsx
@@ -1,44 +1,0 @@
-import { NormalText } from "@/components/Organisms/Static/NormalText";
-
-export const Eligibility = () => (
-  <section>
-    <NormalText>
-      The Site is intended for users who are of legal age. All users who are
-      minors in the jurisdiction in which they reside (generally under the age
-      of 18) must have the permission of, and be directly supervised by, their
-      parent or guardian to use the Site. If you are a minor, you must have your
-      parent or guardian read and agree to these Terms prior to you using the
-      Site.
-    </NormalText>
-    <NormalText>
-      The information provided on the Site is not intended for distribution to
-      or use by any person or entity in any jurisdiction or country where such
-      distribution or use would be contrary to law or regulation or which would
-      subject us to any registration requirement within such jurisdiction or
-      country. Accordingly, those persons who choose to access the Site from
-      other locations do so on their own initiative and are solely responsible
-      for compliance with local laws, if and to the extent local laws are
-      applicable.
-    </NormalText>
-    <NormalText>
-      Additionally, you also represent and warrant that you are not a citizen or
-      resident of a state, country, territory or other jurisdiction that is
-      embargoed by the United States or where your use of the Site or the
-      Services would be illegal or otherwise violate any applicable law.
-      Specifically, you represent that you are not located in, organized in, or
-      a resident of Cuba, Iran, Syria, North Korea, Russia, Crimea, Donetsk,
-      Luhansk, Afghanistan, Balkans, Belarus, Burman, Central African Republic,
-      Congo, Ethiopia, Hong Kong, Iraq, Libya, Lebanon, Nicaragua, Somalia,
-      Sudan and Darfur, South Sudan, Ukraine, Venezuela, Yemen, Zimbabwe or any
-      other jurisdiction where the applicable law prohibits you from accessing
-      or using the Services; and you represent that you are not named in the
-      Office of Foreign Asset Control of the U.S. Department of the Treasury’s
-      Specially Designated and Blocked Persons List.
-    </NormalText>
-    <NormalText>
-      By accessing or using the Site, you agree that you are solely and entirely
-      responsible for compliance with all laws and regulations that may apply to
-      you.
-    </NormalText>
-  </section>
-);
