--- frontend/apps/picasso/components/Organisms/Static/TermsOfUse/EntireAgreement.tsx
@@ -1,16 +1,0 @@
-import { NormalText } from "../NormalText";
-
-export const EntireAgreement = () => (
-  <section>
-    <NormalText>
-      These Terms and any policies or operating rules posted by us on the
-      Services constitute the entire agreement and understanding between you and
-      us and govern your access and use of the Site and Services, superseding
-      any prior or contemporaneous agreements, communications, and proposals,
-      whether oral or written, between you and us (including, but not limited
-      to, any prior versions of these Terms). Any failure by us to exercise or
-      enforce any right or provision of these Terms shall not constitute a
-      waiver of such right or provision.
-    </NormalText>
-  </section>
-);
