--- frontend/apps/picasso/components/Organisms/Static/TermsOfUse/ForwardLookingStatements.tsx
@@ -1,36 +1,0 @@
-import { NormalText } from "../NormalText";
-
-export const ForwardLookingStatements = () => (
-  <section>
-    <NormalText>
-      This information contains “forward-looking statements.” These statements,
-      identified by words such as “plan,” “anticipate,” “believe,” “estimate,”
-      “should,” “expect,” “will,” “can,” and similar future-looking expressions
-      include our expectations and objectives regarding our future operating
-      results and business strategy. Forward-looking statements involve known
-      and unknown risks, uncertainties, assumptions and other factors that may
-      cause the actual results, performance or achievements of Composable and
-      its affiliated entities or related projects to be materially different
-      from any future results, performance or achievements expressed or implied
-      by the forward-looking statements. Such factors include, among others,
-      general business, economic, competitive, political and social
-      uncertainties; dependence on commercial product interest; as well as
-      regulatory or legal changes and uncertainty. Forward-looking statements
-      are based on a number of material factors and assumptions, economic
-      conditions in the near to medium future, the average cost of the Company’s
-      offerings compared to traditional offerings, fluctuations or changes to
-      the tax and other regulatory requirements regarding DeFi and the industry
-      as a whole. While the Company considers these facts and assumptions to be
-      reasonably based on information currently available to it, these
-      assumptions may prove to be incorrect. Actual results may vary from such
-      forward-looking information for a variety of reasons, including but not
-      limited to risks and uncertainties known and unknown by the Company.
-      Because forward-looking statements relate to the future, they are subject
-      to inherent uncertainties, risks and changes in circumstances that are
-      difficult to predict and many of which are outside of our control. The
-      Company’s actual results and conditions may differ materially from those
-      indicated in the forward-looking statements. Therefore, you should not
-      rely on any of these forward-looking statements.
-    </NormalText>
-  </section>
-);
