--- frontend/apps/picasso/components/Organisms/Static/TermsOfUse/GoverningLawAndVenue.tsx
@@ -1,12 +1,0 @@
-import { NormalText } from "../NormalText";
-
-export const GoverningLawAndVenue = () => (
-  <section>
-    <NormalText>
-      These Terms and any separate agreements whereby we provide you Services
-      shall be governed by and construed in accordance with the laws of the
-      Saint Lucia. Any dispute between the Parties that is not subject to
-      arbitration will be resolved in Saint Lucia.
-    </NormalText>
-  </section>
-);
