--- frontend/apps/picasso/components/Organisms/Static/TermsOfUse/Indemnification.tsx
@@ -1,26 +1,0 @@
-import { NormalText } from "@/components/Organisms/Static/NormalText";
-
-export const Indemnification = () => (
-  <section>
-    <NormalText>
-      You hereby agree to defend, indemnify, and hold Composable harmless from
-      and against any loss, damage, liability, claim, or demand, including
-      reasonable attorneys’ fees and expenses, made by any third party due to or
-      arising out of: (a) your access and use of the Site and Services; (b) your
-      breach or alleged breach of these Terms; (c) any breach of your
-      representations and warranties set forth in these Terms; (d) anything you
-      contribute to the Services (e) your misuse of the Services, or any smart
-      contract and/or script related thereto, (f) your violation of the rights
-      of a third party, including but not limited to intellectual property
-      rights, publicity, confidentiality, property, or privacy rights (g) any
-      overt harmful act toward any other user of the Services with whom you
-      connected via the Services; or (h) your violation of any laws, rules,
-      regulations, codes, statutes, ordinances, or orders of any governmental or
-      quasi-governmental authorities. Notwithstanding the foregoing, we reserve
-      the right, at your expense, to assume the exclusive defense and control of
-      any matter for which you are required to indemnify us, and you agree to
-      cooperate, at your expense, with our defense of such claims. You will not
-      in any event settle any claim without our prior written consent.
-    </NormalText>
-  </section>
-);
