--- frontend/apps/picasso/components/Organisms/Static/TermsOfUse/IndemnificationAndFullRelease.tsx
@@ -1,24 +1,0 @@
-import { NormalText } from "@/components/Organisms/Static/NormalText";
-
-export const IndemnificationAndFullRelease = () => (
-  <section>
-    <NormalText>
-      You agree to hold harmless, release, defend, and indemnify Composable and
-      its officers, directors, employees, contractors, agents, affiliates, and
-      subsidiaries from and against all claims, damages, obligations, losses,
-      liabilities, costs, and expenses arising from: (a) your access and use of
-      the Site; (b) your violation of any term or condition of these Terms, the
-      right of any third party, or any other applicable law, rule, or
-      regulation; and (c) any other party&apos;s access and use of the Site with
-      your assistance or using any device or account that you own or control.
-    </NormalText>
-    <NormalText>
-      You likewise expressly agree that you assume all risks in connection with
-      your access and use or interaction with the Site, our Services, and/or the
-      Composable protocols, as applicable. You further expressly waive and
-      release us from any and all liability, claims, causes of action, or
-      damages arising from or in any way relating to your use of or interaction
-      with the Site, our Services, and/or the Composable protocols.
-    </NormalText>
-  </section>
-);
