--- frontend/apps/picasso/components/Organisms/Static/TermsOfUse/InformationalResource.tsx
@@ -1,24 +1,0 @@
-import { NormalText } from "../NormalText";
-
-export const InformationalResource = () => (
-  <section>
-    <NormalText>
-      All information, including graphs, charts, tokenomics, projections, and
-      other data, provided in connection with your access to the Site and the
-      Services are for general informational purposes only and subject to change
-      at the sole discretion of Composable. Composable provides resources about
-      the fundamentals of the Composable protocol or system, which seeks to
-      build a trustless infrastructure for DeFi. This information is not
-      intended to be comprehensive or address all aspects of the protocol. You
-      should not take, or refrain from taking, any action based on any
-      information contained on the Site or any other information that we make
-      available at any time, including blog posts, data, articles, links to
-      third-party content, news feeds, tutorials, tweets, and videos.
-    </NormalText>
-    <NormalText>
-      The materials appearing in the Site could include technical,
-      typographical, or photographic errors. Composable does not warrant that
-      any of the materials on its website are accurate, complete or current.
-    </NormalText>
-  </section>
-);
