--- frontend/apps/picasso/components/Organisms/Static/TermsOfUse/IntellectualPropertyRights.tsx
@@ -1,23 +1,0 @@
-import { NormalText } from "../NormalText";
-
-export const IntellectualPropertyRights = () => (
-  <section>
-    <NormalText>
-      Unless otherwise indicated, the Site is our proprietary property and all
-      source code, databases, functionality, software, website designs, audio,
-      video, text, photographs, and graphics on the Site (collectively, the
-      “Content”) and the trademarks, service marks, and logos contained therein
-      (the “Marks”) are owned or controlled by us or licensed to us, and are
-      protected by copyright and trademark laws and various other intellectual
-      property rights and unfair competition laws of the applicable
-      jurisdiction, international copyright laws, and international conventions.
-      The Content and the Marks are provided on the Site “AS IS” for your
-      information and personal use only. Except as expressly provided in these
-      Terms, no part of the Site and no Content or Marks may be copied,
-      reproduced, aggregated, republished, uploaded, posted, publicly displayed,
-      encoded, translated, transmitted, distributed, sold, licensed, or
-      otherwise exploited for any commercial purpose whatsoever, without our
-      express prior written permission.
-    </NormalText>
-  </section>
-);
