--- frontend/apps/picasso/components/Organisms/Static/TermsOfUse/LimitationOfLiability.tsx
@@ -1,40 +1,0 @@
-import { NormalText } from "@/components/Organisms/Static/NormalText";
-
-export const LimitationOfLiability = () => (
-  <section>
-    <NormalText>
-      Under no circumstances shall Composable or any of its officers, directors,
-      employees, contractors, agents, affiliates, or subsidiaries be liable for
-      any indirect, punitive, incidental, special, consequential, or exemplary
-      damages, including (but not limited to) damages for loss of profits,
-      goodwill, use, data, or other intangible property, arising out of or
-      relating to any access or use of the Site, nor will Composable be
-      responsible for any damage, loss, or injury resulting from hacking,
-      tampering, or other unauthorized access or use of the Site or the
-      information contained within it. Composable assumes no liability or
-      responsibility for any: (a) errors, mistakes, or inaccuracies of content;
-      (b) personal injury or property damage, of any nature whatsoever,
-      resulting from any access or use of the Site; (c) unauthorized access or
-      use of any secure server or database in our control, or the use of any
-      information or data stored therein; (d) interruption or cessation of
-      function related to the Site; (e) bugs, viruses, trojan horses, or the
-      like that may be transmitted to or through the Site; (f) errors or
-      omissions in, or loss or damage incurred as a result of the use of, any
-      content made available through the Site; and (g) the defamatory,
-      offensive, or illegal conduct of any third party. Under no circumstances
-      shall Composable or any of its officers, directors, employees,
-      contractors, agents, affiliates, or subsidiaries be liable to you for any
-      claims, proceedings, liabilities, obligations, damages, losses, or costs
-      in an amount exceeding the amount you paid to us in exchange for access to
-      and use of the Site, or USD$50.00, whichever is greater. This limitation
-      of liability applies regardless of whether the alleged liability is based
-      on contract, tort, negligence, strict liability, or any other basis, and
-      even if we have been advised of the possibility of such liability. Some
-      jurisdictions do not allow the exclusion of certain warranties or the
-      limitation or exclusion of certain liabilities and damages. Accordingly,
-      some of the disclaimers and limitations set forth in these Terms may not
-      apply to you. This limitation of liability shall apply to the fullest
-      extent permitted by law.
-    </NormalText>
-  </section>
-);
