--- frontend/apps/picasso/components/Organisms/Static/TermsOfUse/NotAnOffering.tsx
@@ -1,29 +1,0 @@
-import { NormalText } from "../NormalText";
-
-export const NotAnOffering = () => (
-  <section>
-    <NormalText>
-      Any information in this Site does not constitute an offer to sell or a
-      solicitation of an offer to purchase securities, assets, including digital
-      assets, or financial instruments by the Company, or to enter into a
-      transaction involving any such security or financial instrument. Such an
-      offer can only be done through a registered or licensed offering or
-      subject to an exemption. The recipient should not rely upon anything
-      within this information in making a decision to participate in the
-      Company’s issuances or to utilize the Company’s technology. The Company is
-      not required to update the information provided and the information is
-      only current as of the date of its release and is subject to change over
-      time.
-    </NormalText>
-    <NormalText>
-      Any information provided in this Site is not intended for distribution to
-      or use by any person or entity in any jurisdiction or country where such
-      distribution or use would be contrary to law or regulation or which would
-      subject us to any registration requirement within such jurisdiction or
-      country. Accordingly, those persons who choose to access the Site from
-      other locations do so on their own initiative and are solely responsible
-      for compliance with local laws, if and to the extent local laws are
-      applicable.
-    </NormalText>
-  </section>
-);
