--- frontend/apps/picasso/components/Organisms/Static/TermsOfUse/NotProfessionalAdvice.tsx
@@ -1,24 +1,0 @@
-import { NormalText } from "@/components/Organisms/Static/NormalText";
-
-export const NotProfessionalAdvice = () => (
-  <section>
-    <NormalText>
-      All information provided by the Site or Services is for informational
-      purposes only and should not be construed as professional advice. You
-      should not take, or refrain from taking, any action based on any
-      information contained in the Site or Services. Before you make any
-      financial, legal, or other decisions involving the Site or Services, you
-      should seek independent professional advice from an individual who is
-      licensed and qualified in the area, subject matter and jurisdiction for
-      which such advice would be appropriate.
-    </NormalText>
-    <NormalText>
-      Composable is not your broker, lawyer, intermediary, agent, or advisor and
-      has no fiduciary relationship or obligation to you regarding any decisions
-      or activities that you have undertaken or will be undertaking when using
-      the Site or the Services. Neither our communications nor any information
-      that we provide to you is intended as, or shall be considered or construed
-      as advice.
-    </NormalText>
-  </section>
-);
