--- frontend/apps/picasso/components/Organisms/Static/TermsOfUse/ReservedRights.tsx
@@ -1,15 +1,0 @@
-import { NormalText } from "@/components/Organisms/Static/NormalText";
-
-export const ReservedRights = () => (
-  <section>
-    <NormalText>
-      Composable reserves the following rights: (a) with or without notice to
-      you, to modify, substitute, eliminate or add to the Site; (b) to review,
-      modify, filter, disable, delete and remove any and all content and
-      information from the Site; and (c) to cooperate with any law enforcement,
-      court or government investigation or order or third party requesting or
-      directing that we disclose information or content or information that you
-      provide.
-    </NormalText>
-  </section>
-);
