--- frontend/apps/picasso/components/Organisms/Static/TermsOfUse/UseCases.tsx
@@ -1,13 +1,0 @@
-import { NormalText } from "@/components/Organisms/Static/NormalText";
-
-export const UseCases = () => (
-  <section>
-    <NormalText>
-      Use cases are provided for illustration purposes only and the Company
-      neither supports nor condones any such uses. Any such uses should be done
-      at the risk of the person making the utilization of the technology and any
-      user agrees to hold harmless and indemnify the Company from and against
-      any losses that may occur as a result of any such uses.
-    </NormalText>
-  </section>
-);
